requirements.txt
fastapi
unicorn
openai
python-dotenv